Modify signup flow
